@@ -1,4 +1,4 @@
-class AddUserIdToTweets < ActiveRecord::Migration
+class AddUserIdToTweets < ActiveRecord::Migration[4.2]
   def change
     add_column :tweets, :user_id, :integer
   end
